backend/requirements.txt
Django>=5.1.5
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.4.0
django-cors-headers>=4.6.0
django-filter>=24.3
django-allauth>=65.4.0
python-decouple>=3.8
requests>=2.32.3
PyPDF2>=3.0.1
dj-database-url>=2.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
python-dotenv>=1.0.1
twilio>=9.10.0